jumping and attacking!
W = jump
Y = attack

notes:
updated the animation drawing method to support "lingering" animations,
such as jump, that may linger on the last frame for some period of time.
currently, the jumping is not collusion based, and that will be changed
eventually (the jump length is also hard-coded to be 2x the animation
duration).
you are allowed to change directions while moving, but the animation is
locked. perhaps lock momentum when you start jumping?
also, i made it so you can't attack in midair. maybe we'll change that -
seeing as the Q already looks like an aerial skill, it might not look
that bad.